HYD-4898 The mdt/ost_next_index counter can be corrupted due to ObjectCache issues.
the updates to the index can be lost because they are not written to the ObjectCache(sic)
just to the database. This means that if something else reads from the ObjectCache and
then saves the values in the index are lost.

The simple solution is to update the ObjectCache when updating the indexes.

 - Update create_filesystem to create and commit the filesystem and mgs before
   creating the filesystem targets. Committing before creation makes the state
   of the database and ObjectCache consistent before we create the targets.
 - Make create_for_volume use the ObjectCache version of the filesystem rather
   than the version on the database.

Change-Id: If897e7dd3d6413bb944d4889b525f76e7aee862b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/16780
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>