HYD-4961  Agent unit test fails intermittently due to string ordering
This is a child issue of HYD-4965, discovered when trying to improve
reliability of agent unit tests before testing with different version of
python (2.7) the default for RH7.2.

There is an intermittent failure in agent unit test test_scan_packages related
to the order of the repo names passed to the --enablerepo parameter of the
mocked out "yum clean" shell command.  The order of these repo's is not
deterministic due to the processing of the unordered dict in
lustre.py:scan_packages() before repo names are passed in commands
to the yum_utils classes.

In order to resolve this issue the repo names are processed in alphabetical
order within scan_packages and when tested alphabetical lists are passed in and
compared.

  - Change test case to provide and expect alphabetic list of repo names
  - Process lists of repos in alphabetical order

Change-Id: I65de8e150cf7b9e46e1bde5afc8533da3eebbc7c
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/17146
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/22679
Tested-by: Chroma Test User