HYD-4994 To support corosync1 and corosync2 with selectable objects
Corosync1 and Corosync2 have different stateful objects. This patch
does the ground work to allow the objects create, lnet, rsyslog, ntp
etc to be selected as part of the profile. This is part of the work
towards allowing for user selected/defined profiles.

The simple method is the profile attributes decided with stateful objects
are created on the host, and agent that for example does/doesn't have
the rsyslog stateful object created will/will-not manage the rsyslog
changes.

We control the profiles and so can control what is and isn't created,
so this solutions provides for creator flexibility for use.

 - Create some RH7 profiles
 - Update profiles with new attributes.
 - Update the code to work with the new data driven object creation.

Change-Id: Ia1fcbddb1e97dd681c7334d2e35b3ec74602b3a8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17792
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>