HYD-5075 Stop asking the provisioner to intall a Lustre client
The provisioner provisions a Lustre client on the client node of a test
cluster if asked to. And our cluster configurations ask it to.
But we never actually end up using that Lustre client and our tests install
a new one on top of the one installed by the provisioner. So let's just
stop wasting the provisioning time doing it, OK?

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: If908c8579f5947c68201d0ee0aea89d6297c072f
Reviewed-on: http://review.whamcloud.com/17206
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>