HYD-5092 Update build processes / scripts to account for new module names
As a result of moving our code to internal Github / Artifactory, we now
need to update our current makefiles and test scripts to account for
changes.

The largest change to account for is Artifactory not supported scoped
modules, which means all the module names need to be updated in scripts.

Ziplocker has also changed as a result of the migration. Some of it's commands
will need to change as well.

Change-Id: I51c1607d0e35196785fe2563abeb0307c18b7075
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/17238
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Tested-by: Jenkins
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>