HYD-5095 Initial 3.0 Help Update
With every release, the technical writer updates the Help to reflect
changes to the product. This is the initial update for Help
needed for the 3.0 release, prior to stabilization testing. 

The Help files consist of a set of HTML and other files that are 
generated by the "Help and Manual" writing tool. The source is in 
XML and the generated HTML file-set result is what the writer 
commits to the IML project. 

The only changes the writer has made for this initial release are 
the rolling of the version to 3.0. (Other files may be updated by
 Help and Manual, these being changes to metadata relating to 
new internal paths, Help project information, etc.)

To confirm this commit was successful, open Help and check 
the Help cover page.

Change-Id: Ib043115ee197d24d7b266e3da4207c5414c223a1
Reviewed-on: http://review.whamcloud.com/17216
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>