HYD-5093 agent deployment script template is string rather than file
Agent http setup endpoint (views.py) on manager http_agent service
populates a script template defined as a multi-line string.  Seeing
as this is executable code to be run by the agent it would be useful
if we could view and change the code as a python file.

The solution chosen is to supply the script templates as a .template
file in chroma-manager/chroma_agent_comms/ and read file in as text
when needing to populate template in views.py.

  - Add agent-bootstrap-script.template file to chroma-manager dir
  - Remove setup_script_template string definition in views.py
  - Read template text from file prior to populating template
  - Specify template path location relative to running file
  - Add template file to build manifest for chroma-manager pkg
  - Add relevant entries in spec file to add template to RPM pkg

Change-Id: I6e51e8ffdd9e0187ab7b7feddb674ff724085cec
Signed-off-by: tanabarr <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/17210
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>