HYD-4867 Test command alerts
Provide a set of basic integration tests for the IML command alerts.
Ensure that the command alerts trigger correctly when a specific set
of commands are executed.

To do this test the lnet configuration state was changed
from up to down or vice versa.The cancel alert test checks
for both the 'Running Command Alert' and 'Cancel Command Alert'
this was done by checking the command alert for the lnet configuration
was a Running Alert after the PUT request and then if this was true
cancel the command to change the lnet state and check if the cancel
command alert was triggered off. The error alert was done by stoping
the agent before the PUT request to change the lnet configuration which
results in a error and the successful alert was simply done by giving the
command and letting it finish and checking the successful command alert
was raised.

 - Add test_command_alerts.py with 3 separatetests to cover
   the required functionality.

Change-Id: Iaa69e9ee1e053a9020fb42a319d350508784c0d1
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/17825
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>