HYD-5090 Build on EL7 as well as EL6
3.0 is required to support the Manager and the Agent on El6 and
EL7, this requires being able to build them on EL6 and El7 builders.

Change the build variables to be generic EL variables rather than EL6
variables i.e change BUILDER_IS_EL6 to BUILDER_IS_EL. Use the PLAT_BUILD_OK
variable to control of a particular parget is built. For example 
PLAT_BUILD_OK evaluates to False for erlang on EL7 causing erlang to
not be built on EL7.

 - Add some useful extra ignores to .gitignore
 - Update the root Makefile to switch on EL rather then EL6 in particular.
 - Update chroma-dependencies Makefiles to selectively build depending
   on the value of PLAT_BUILD_OK
 - Extra piece of debug in chroma-manager/Makefile

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I7314e98d8e4f166f5c83ed22fb279573c5e0c43e
Reviewed-on: http://review.whamcloud.com/17283
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>