HYD-5105 Don't re-use existing RPM variables
The addition of changes to build EL7 and EL6 require the use of the %{dist}
in the release of our packages to indicate which distro release they are for.
This has unearthed a situation where existing variables such as %{release} are 
reused causing the Release portion of the package name to be duplicated.

This is resolved by changing the name of the variable release to be 
package_release.

 - Update the root Makefile and Spec file for agent, diagnostics, manager and
   the simulator to use package_release in place of release.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I576ccf7eea34f1a417a358a687fedda8e1650373
Reviewed-on: http://review.whamcloud.com/17284
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>