HYD-5112 Cannot have .py files with not-valid-python code in them
Having a file ending with .py that doesn't really have Python code in it
fails during the compilation/validation phase of an RPM build on EL7. 
The IML validation has a "test" file with intentionally non-Python
content to test the effectiveness of a syntax checking tool, this file
causes the RPM validation to fail.

In order to make the IML tests compatible with the RPM validation on EL7
the file must be renamed to a non .py name and then the tests updated to
temporarily give the file a .py extension during the test.

 - Rename junk.py to junk.not-really-py
 - Update test_manager.py to rename the file to junk.py during the test.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Id7f2c7e2fd11dc8de2d0eb5cb2403b63c7409624
Reviewed-on: http://review.whamcloud.com/17291
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>