HYD-5134 Build pyparsing instead of pulling from Epel
pyparsing it is not available in Epel7 requiring us to
build it from source. To keep the build process as consistant
as possible across different builds the build process should
build pyparsing from source for EL6 and EL7.

Add pyparsing/Makefile to build from source and remove the
binary rpm's from the Epel Makefile.

 - Replace pyparsing binary rpm with the source rpm in 
   chroma-externals
 - Add Makefile to build the source
 - Remove the binary from the Epel Makefile.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Iaad2481a407404586c6745e4a1ba1986c7c4a4ec
Reviewed-on: http://review.whamcloud.com/17300
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Tested-by: Chris Gearing <chris.gearing@intel.com>