HYD-5132 Build Django-south instead of pulling from Epel
Django-south it is not available in Epel7 requiring us to
build it from source. To keep the build process as consistant
as possible across different builds the build process should
build Django-south from source for EL6 and EL7.

Add Django-south/Makefile to build from source and remove the
binary rpm's from the Epel Makefile.

 - Replace Django-south binary rpm with the source rpm in 
   chroma-externals
 - Add Makefile to build the source
 - Remove the binary from the Epel Makefile.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I1854116fafce5084e7afddd40a8c18c8a85f947c
Reviewed-on: http://review.whamcloud.com/17298
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>