HYD-5144 Simulator doesn't always work after HYD-4255 landed.
The simulator was broken after HYD-4255 landed when it is run from
the command line rather than the CI system.

The problem is that the nice classes (we really wanted structs) used
so represent Corosync and Pacemaker are saved to disk as dicts and so
when they return are still dicts.

Solve this by casting them back to the classes after they are loaded.

 - Base the Pacemaker/Corosync classes on a more generic solution that
   can be called ot convert dicts that are really 'structure' back into
   those stuctures
 - Call this new convert_dict function after reloading the dict state.

Change-Id: If05b62a81491abea1e4b392c47c2e957a8c5a4f0
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17330
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>