HYD-5133 Build django-picklefield instead of pulling from Epel
django-picklefield it is not available in Epel7 requiring us to
build it from source. To keep the build process as consistant
as possible across different builds the build process should
build django-picklefield from source for EL6 and EL7.

Add django-picklefield/Makefile to build from source and remove the
binary rpm's from the Epel Makefile.

 - Replace django-picklefield binary rpm with the source rpm in
   chroma-externals
 - Add Makefile to build the source
 - Remove the binary from the Epel Makefile.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I9f6423b4023844aab525b98aed7cf74bacdc1b30
Reviewed-on: http://review.whamcloud.com/17299
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>