HYD-5090 Dependency updates for EL7
From EPEL:
- pcapy updated
- python-daemon updated
- python-importlib removed for EL7; now included in Python 2.7
- python-lockfile updated
- python-meld updated
- python-ordereddict updated
- v8 updated only in packaging
- nodejs updated
- libuv updated
- c-ares19 removed for EL7
- pdsh removed for both EL6 and EL7 as we don't use/depend on it

Update nginx to version 1.8.0 since we needed to also get the EL7
build and 1.8.0 is what was available currently.

Gunicorn got moved to the extras repo/channel, so rather than asking
users to add yet another repo/channel, let's just distribute it our-
selves, the same way we do with EPEL dependencies.

When done with them, remove any build-time dependencies that were
installed so that packages could be built.  This keeps the builder
clean for the next build.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I03089b28328df6862e5425d84aed379543964731
Reviewed-on: http://review.whamcloud.com/17286
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>