HYD-5150 SetupWorkerJob spelt SetupWorkertJob in HYD-4255
SetupWorkerJob is spelt SetupWorkertJob in HYD-4255. Changing
it now means we have no misspellings in the field. It is also
the case that errors exist in the table definitions for the
migrations - the migrations are OK but an error exists due to
a manual change.

chroma_core.ConfigureCorosyncJob should be chroma_core.configurecorosyncjob

 - Update SetupWorkertJob to SetupWorkerJob and amend migration
 - Update chroma_core.ConfigureCorosyncJob in model definitions in migration

Change-Id: I00d7d6acf4b4210b197acd5f177d62702d6c81fe
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17613
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>