HYD-4455 Test failure: integration, existing_filesystem_configuration
Cause is that the update devices step for each host is serialised and
so takes longer than the test expects, zfs can take long time to scan
devices due to the need to import currently exported zfs pools. On a
big system detection could take hours, days! - so increasing the
timeout is not a solution.

The solution is to parallelize it with threads within the step. The
code being threaded already operates a lot on multiple threads so
can be presumed thread safe.

 - Make the UpdateDevicesStep operate on multiple hosts in parallel.
 - Change the users to UpdateDevicesStep to provide a list of hosts to
   operate on.

Change-Id: I2af8a45c0c911803d2a58be03942aa7171df6515
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17388
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>