HYD-4307 Use platform_info script to derive platform values
Instead of hardcoding the platform values in
chroma/chroma-dependencies/include/Makefile.meta make use
of the include/platform_info script to determine them.

This will increase our robustness to adapting to platform
changes in the future.

With this patch, any chroma-dependency/* that you want built
on a given platform needs to have a .<platform> (i.e. el-6,
el-7) added to it's directory.

 - Add an empty .el6 and .el7 file to all chroma-dependencies
   that require building on those platforms. Which is all other
   than erlang that builds only on .el6
 - Revert previous changes to Makefile.meta so that it again uses
   the platform_info script
 - Update the platform_info script to be compatible the latest Ohai
 - Add trollop.rb because it is required by the platform_info script.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: If93125f16aee313fd11cf1d3b250b7925e297bae
Reviewed-on: http://review.whamcloud.com/14163
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>