HYD-5184 Build python-anyjson from EPEL SRPM
With python-anyjson we are not following our principle of putting
trust first in upstream RPMs before deciding to roll our own.

This patch rectifies that by grabbing the appropriate upstream
SRPM from EPEL and building that.

We in fact could and should just be grabbing this from EPEL in binary
form however it needs to be installed as a dependency to build kombu
and Makefile.fetch-from-repo does not support that functionality.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ie240363389beff7adb4a2dd7d38f1974620d12af
Reviewed-on: http://review.whamcloud.com/17435
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>