HYD-5182 Force %dist to be .el7 on CentOS
Because one of the changes that CentOS makes is to redefine the %dist
macro from .el7 to .el7.centos, we need to force it back to .el7
during our package builds.

So when we build our own RPMs strip any .centos out of the %dist macro.
This is done for chroma-dependencies as well the chroma RPMs.

chroma-externals: 17279/2
Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I310d8201beb10ecb02d7737c348a568c64d0438b
Reviewed-on: http://review.whamcloud.com/17479
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>