HYD-5147 Create a class to manage linux services states
Provide a way to manage linux services state, to make issuing ‘service’
command executions more defensive and robust.

To do this a Parent class and child class was implemented.
the parent class contains all the retry and check logic with
the child class having the actual commands in it.
It works by the parent class calling a create method which
returns a child class object that contains the commands applicable
with the linux version that is being used (RH6).
For stopping,starting and restarting the service, checking the
running state after a certain amount of time was done to check
the service state stayed the same after the command execution
returned successful.

 - Add service_control.py with 2 classes to cover
   the required functionality.

Change-Id: Ic396efa224d140f90ecc278995ce39f01ae50e92
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/17480
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>