HYD-5191 Add simplejson back into chroma-externals for use with EL7
HYD-5018 removed Simplejson from the Agent, Manager and from chroma-externals.
Removal of simplejson from the Agent code makes sense the removal from
chroma-externals was a mistake because python-django requires simplejson.

Generally minimizing the dependencies of a product has many advantages mostly
in terms of management and maintenance costs.

The removal of simplejson from the Agent and Manager reduces the number of
externals directly by IML and provides for a more consistent use of json. It
carries no risk because we completely understand the usages of the json library
within the IML code.

Removing simplejson from django carries more risk because we do not understand
all the usages of json within django and we do not know why the upstream
maintainer explicitly made django use simplejson.

simplejson is part of core-0 for EL6 and so we only need to provide the simplejson
rpm for EL7, the version added is 3.3.3-1.el7.x86_64 sourced from the EPEL 7 repo.

 - Add simple json from epel7 to chroma-externals
 - Add python-simplejson-3.3.3-1.el7.x86_64.rpm to the EPEL RPMS list for EL7

Change-Id: I3596af4fe398fe005e9b0e4417b7486a6ed736cf
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17517
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>