HYD-5195 Selenium: ZeroDivisionError: float division
Client connections are calculated has the total number of connections
divided by the number of mdts. Durring the creating of a filesystem and
the deletion of a filesystem the number of mdts can actually be zero.
If filesystem information is requested at this time the calculation of
the client connections fails.

Update the filesystems endpoint to deal with this corner case by presuming
that a filesystem with 0 mdts will have 0 client connections.

 - Update filesystem.py to detect when there a no mdts present and then
   presume client count is 0

Change-Id: I974e3cafc4f42124d5248c8c4b91eeaf1e7d614b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17528
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>