HYD-5178 Test Failure: Long polling test can fail.
Long polling tests can fail because the number of updates of the
database is un predictable. As well as needing to fix the
unpredictability the quality of the debug needs to be improved.

Step one is to add addtional information when long polling
integration tests fail. Then information provided for debug is not
useful. Maximizing the information will allow the failures to be more
easily debug.

Extend the wait_until_true function to have a custom error message
including the ability to make the message callable so that it can
be created after the error rather than before the error. This allows
post error information to be included.

Change the resource_manager.py to only update the NetworkInterface
and Nid records if those records change. This is more predictable
and as a knock on effect makes the test_host_long_polling (used to
 be incorrectly called test_alert_long_polling) predictable.

Make use of the functionality in the long polling tests.

 - Update wait_until_true to allow flexible error messages
 - Update test_lnet_long_polling and test_alert_long_polling to
   use the new wait_until_true functions.
 - Rename test_alert_long_polling to test_host_long_polling
 - Update resource_manager.py to check for changes to
   NetworkInterface and Nid before writing to the database.

Change-Id: I6d6df9c17c3e16e8d8c470d9499d208e82a63156
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17587
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>