HYD-5218 _fetch_help repeatedly requests help.
The _fetch help function is very useful but in the case of a repetitive
failure within the same test session repeatedly requests help and as each
request takes 15 minutes the test can take days to run.

fetch_help should be updated to request help only once for each type of
failure. As changes were being made moved _fetch_help up to classes in
the hierarchy. _fetch_help was being referenced from the parent of the
class that it was defined in.

 - Add a list to fetch_help that allows it track and issue help only
   once for each failure.
 - Remove an unrequired _fetch_help call from the lnet tests.
 - Moved _fetch_help from ChromaIntegrationTestCase to UtilityTestCase.

Change-Id: Ie8fbba2521e7781f46a839d08e5dd61b4e476c0b
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17586
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>