HYD-5191 Include simplejson in the requirements exclusion
When the previous patch for this ticket landed, it did not
resurrect the simplejson tarball for installing with pip
in the manager's virtualenv.

But clearly this is not a requirement because the El6 build
finished without needing a simplejson tarball.

So the solution here is to simply exempt the simplejson
module from being installed in the manager's virtualenv
which is only used to build the docs.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ia47d2f33a3e4a896d59c870a9ee4c4985fcaa2b4
Reviewed-on: http://review.whamcloud.com/17711
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>