HYD-5247 Stray threads after test: ['DatabaseChangedThread'] error
One of the unit tests, TestDetectedFSTransitions.test_forget, creates an
instance of the HttpAgent Service. Because of the way long polling is enabled
this starts the long polling database hooks. We don't want these running in
units tests and so as a refactor moving the enabling to the run method of
ChromaService will prevent this as the unit tests do not run the Service.

To do this the start method has to be implemented by the ChromaService job and
the descendants changed to call the superclass version. To keep the
implementation symmetrical the descendants also call the superclass version of
stop.

  - Modified the following files to call start and stop in the base
    ChromaService class.
    chroma_core/services/corosync/__init__.py
    chroma_core/services/http_agent/__init__.py
    chroma_core/services/job_scheduler/__init__.py
    chroma_core/services/lustre_audit/__init__.py
    chroma_core/services/pinger.py
    chroma_core/services/pinger_client.py
    chroma_core/services/plugin_runner/__init__.py
    chroma_core/services/power_control/__init__.py
    chroma_core/services/stats.py
    chroma_core/services/syslog/__init__.py
  - Modified the base class to enable long polling in run rather than
    at creation. chroma_core/services/__init__.py

Change-Id: I220d0faffbdfa1201e8a108e29a35aa28669d239
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17717
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>