HYD-5090 Clean up Makefile EL7 reorganization
The reoranization of the EPEL chroma-dependency Makefile
did not pay enough attention to keeping the various variables
populated with only values respecitve of the release that was
being built.

Also, because the python-daemon package is different versions
in EPEL 6 and EPEL 7, the tarball for the EPEL 7 version was
missing.

This patch cleans up the Makefile and adds that missing tarball.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I868daad59fb55bed1e38762055fa40919c0a12c0
Reviewed-on: http://review.whamcloud.com/17725
Tested-by: Jenkins
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>