HYD-5300 Upgrade the OS as part of the lustre upgrade
Upgrades between IML version often require the OS to be upgraded
prior to IML. Previously this was ignored as part of the upgrade
process and the previous IML version was hacked to be able to install
on a previous version. The framework needs to be expanded to update
the OS before upgrading iml.

Define TEST_DISTRIBUTION and UPGRADE_DISTRIBUTION to be the start and
end version for IML upgrade tests.

 - Upgrade defaults.sh to include new definitions. These can be overridden
   for matrix jobs.
 - Add a new file with a function that upgrades the OS.
 - Include and call the new os upgrade function from run_tests for both
   upgrade suites.
 - Handle TEST_DISTRIBUTION appropriate as it's value has to be overridden
   for upgrade tests.

Change-Id: If6e51b15d2265ea051e5084fffb264e52814a5a3
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17867
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>