HYD-5334 Enable the CorosyncNoPeersAlert
As part of HYD-4255 and new Alert was created the would inform the
user when a HA node had no peers. This is a common case when the
network has not been configured correctly. For test reasons this alert
was not enabled in the HYD-4255 patch.

This patch enables the Alert with no other functional changes and added
specific tests for the alert.

 - Enable alert in Corosync service.
 - Add and update integration tests.

Change-Id: Ib8a3f83069b072cfceaeee3079f63a04ae848570
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17903
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>