HYD-5298 Fix bit rot in code to use rhel for distro in matrix based tests
Two of our test suites, the upgrade os and the installation and upgrade suites,
are run both on CentOS and Rhel driven by a matrix from Jenkins. Their test
scripts were each checking which slave they were on, and for rhel, were running
a sed on provisioner_input.json so rhel was provisioned rather than centos.
However, since rhel testing has been turned off on master until it can support
rhel6.7, this line of code has suffered bit rot twice over. First, since we
switched to using TEST_DISTRIBUTION instead of explicit versions in our json,
a sed on the file does nothing. Secondly, when there was some refactoring done
to reduce duplicate code for provisioning, the creation of
provisioner_input.json was moved until later in the code so it's trying to sed
a file that doesn't exist yet, resulting in a halting error.

To fix this, we simply need to adjust the TEST_DISTRIBUTION variable to be for
the proper distro, rather than mucking with json files, and that will be
propagated to later code that already provisions based on the contents of that
variable.

  - Replace sed on provisioner_input.json with TEST_DISTRIBUTION in each
    chroma-manager/tests/framework/integration/installation_and_upgrade/jenkins_steps/main
    and chroma-manager/tests/framework/upgrade_os/jenkins_steps/main
  - One line unrelated minor fix to restore previous behavior of printing
    provisioner_output.json which makes it easier to debug issues related to
    provisioning.

Change-Id: I2d0ff02cdad2387dc99547965ed532628c178ec1
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/17915
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>