HYD-5311 Clear out ignored files during CI build
During CI builds we clear out ignored files in chroma-manager by hand.
This approach is brittle and one needs to remember to update this list.
It also has the issue of newer branches leaving cruft for older ones.
This causes older branches to have to remove newer branches code in a
curated list.

Instead of that, we can try removing ignored files / dirs using git
clean -ndX $dirname and rm rf the results.

  - Remove ignored files dirs in chroma-manager.
  - Remove curated list.

Change-Id: Ic9a0daa969f783a6b54b26b009c74d0089244a2c
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/17975
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>