HYD-5317 Commands running when Managed exits unexpectedly are left 'running'
Each Command not has an alert associated with it. For a running command this
is a CommandRunningAlert. These alerts changed to CommandCancelledAlert,
CommandErroredAlert or CommandSuccessfulAlert.

However if the JobScheduler crashes the commands today are just set to cancelled
without the corresponding alert being changed - we end up with CommandRunningAlerts
for commands that are not running.

To fix this change Run in the JobScheduler service needs to use the Command completed
method rather than just updating the database directly.

 - Update __init__.py to use Command.completed rather than a direct database modification.

Change-Id: I4e612cf3f2e371c30676d9ce68e61c7d3f7815a0
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/17996
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>