HYD-5324 mcast detection has a hole which can lead to cluster detection failure
In the mcast detection the nodes listen for the other node on ring1
to find out the mcast number. If they don't hear anything they start
talking using the talker thread.

However the talker thread is stopped and then corosync starts - in the
time between one stopping and the other starting the other node can go
through the detection process and because is doesn't hear the other
node - because it is post talker thread but pre corosync starting - can
choose another mcast post.

The solution is to have the talker thread keep talking until corosync starts.

This patch takes the current talker class and uses it in a way that leaves
it running after the detection until corosync starts when the node running
is the node that 'picked' the port.

 - Move TalkerThread into a file on its own
 - Add some tests for the TalkerThread
 - Format error in log.py fixed
 - Corosync2 updated with some debug

Change-Id: I4697de1778de020e3242dc8637e3b1408ed8e0f9
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18063
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>