HYD-5328 Tighten up el6.4 distro substitutions
Due to our inability to keep our Jenkins configuration current with distros,
we have code in IML and it's tests that changes el6* distro definitions to
the el6.4 that we have maintained in Jenkins for a long time.

Some of these are overly aggressive in what they change and this is causing
el7 builds to have their chroma repos changed to el6.4.

So tighten up when these el6.4 substitutions happen so that they are targeted
and precise.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I5f1dc6435f9010d71377c83cefefcc8cb5195406
Reviewed-on: http://review.whamcloud.com/18053
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>