HYD-5341 Adding servers with root password authentication fails
Adding servers using passwords rather than keys failed because a
parameter had been added to create_host_ssh in the jobscheduler
and presumably not to the caller create_host_ssh in the jobscheduler
client. The error was not found because of lack of testing but
also because of the use of optional arguments. Optional make it
impossible to detect incorrect parameters even at runtime.

Removal of the additional variable and removal of the optional
parameters resolves the issue. We still lack testing to capture 
the use of passwords rather than keys but enabling that would 
be quite a significant piece of work beyond the scope of this 
ticket.

 - Remove additional and unused argument target_state
 - Removal optionality from root_pw, pkey and pkey_pw

Change-Id: I929679c3f31550e473f6facbf0f45f312882c00e
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/18086
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>